[쇼핑 주문] 채채 3, 4단계 미션 제출합니다. (#91)
* feat: suspend 적용 및 updateCartItem 함수 분리

* refactor: id값 Long 타입으로 변경

* chore: 인증 id 값 변경

* refactor: clickCheckAll -> clickToggleAll 로 함수명 변경

* refactor: updateCartItem 함수 분리에 따른 장바구니 추가 삭제 로직 수정

* feat: 쿠폰 페이지 view 구현

* feat: OrderFragment 기본 세팅 구현

* feat: 쿠폰 api 데이터 및 도메인 레이어 구현

* fix: CouponDto 클래스 변경

* feat: 주문 fragment에 선택된 장바구니를 넘기는 로직 구현

* feat: 쿠폰을 보여주는 recyclerview 구현

* feat: 쿠폰 적용 로직 구현

* feat: 최소주문 조건에 따른 visibility 설정 구현

* feat: 쿠폰을 최대 1개만 적용할 수 있는 로직 적용

* feat: 쿠폰과 주문 에러 헨들링

* feat: cartItemIds를 long 타입으로 변경 및 최종 결제 로직 구현

* test: 코드 수정에 따른 테스트 코드 수정

* refactor: test용  repository 네이밍 수정

* fix: 쿠폰 공통 로직 정리 및 잘못된 로직 수정

* refactor: 필요없는 repository 주입 삭제

* test: 쿠폰 로직 test 코드 작성

* fix: 장바구니 추가 로직 오류 수정

* refactor: 에러 핸들링 추가

* refactor: runCatching을 사용하여 result 반환하도록 수정

* refactor: 쿠폰의 필수 정보가 없을 경우 기본값으로 바꾸는 것이 아닌 에러가 나도록 수정

* refactor: dao suspend fun으로 설정

* refactor: ProductListViewModel 에서 CoroutineExceptionHandler 을 사용한 에러 핸들링

* refactor: ShoppingCart 에러 핸들링 및 예외를 세분화

* refactor: BaseViewModel을 사용하여 공통된 에러 이벤트 처리

* refactor: ProductDetailViewModel coroutineExceptionHandler을 사용하도록 수정

* refactor: viewModelScope.launch 를 사용하도록 변경

* refactor: coupon 전략 패턴을 사용하여 적용

* test: test 코드 수정

+++ app/src/main/java/woowacourse/shopping/data/remote/dto/cart/CartOrderRequest.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class CartOrderRequest(
     @SerializedName("cartItemIds")
-    val cartItemIds: List<Int>,
+    val cartItemIds: List<Long>,
 )
